Register and Checkout function
