+++ demoProject/src/main/java/Controller/LoginController.java
@@ -1,7 +1,6 @@
 package Controller;
 
-import DAO.LoginDAO;
-import Services.Connect;
+import DAO.UserDAO;
 import nhom26.User;
 
 import javax.servlet.ServletException;
@@ -11,8 +10,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
-import java.sql.*;
-import java.util.Date;
 
 @WebServlet(name = "LoginController", value = "/login")
 public class LoginController extends HttpServlet {
@@ -23,8 +20,8 @@ public class LoginController extends HttpServlet {
         resp.setCharacterEncoding("UTF-8");
         String email = req.getParameter("email");
         String password = req.getParameter("password");
-        LoginDAO loginDAO = new LoginDAO();
-        User user = loginDAO.getUserByEmailAndPass(email, password);
+        UserDAO userDAO = new UserDAO();
+        User user =userDAO.getUserByEmailAndPass(email, password);
         if(user != null){
             HttpSession session = req.getSession();
             session.setAttribute("user", user);
